Refactor ron swanson meme to generator
